Move HDF5 dataset to single file
